+++ lib/keisan/ast/time.rb
@@ -1,6 +1,10 @@
+require_relative "date_time_methods"
+
 module Keisan
   module AST
     class Time < ConstantLiteral
+      include DateTimeMethods
+
       attr_reader :time
 
       def initialize(time)
@@ -11,48 +15,8 @@ module Keisan
         time
       end
 
-      def +(other)
-        other = other.to_node
-        case other
-        when Number
-          self.class.new(value + other.value)
-        else
-          super
-        end
-      end
-
       def to_s
         value.strftime("%Y-%m-%d %H:%M:%S")
-      end
-
-      def >(other)
-        other = other.to_node
-        other.is_a?(AST::Time) ? Boolean.new(value > other.value) : super
-      end
-
-      def >=(other)
-        other = other.to_node
-        other.is_a?(AST::Time) ? Boolean.new(value >= other.value) : super
-      end
-
-      def <(other)
-        other = other.to_node
-        other.is_a?(AST::Time) ? Boolean.new(value < other.value) : super
-      end
-
-      def <=(other)
-        other = other.to_node
-        other.is_a?(AST::Time) ? Boolean.new(value <= other.value) : super
-      end
-
-      def equal(other)
-        other = other.to_node
-        other.is_a?(AST::Time) ? Boolean.new(value == other.value) : super
-      end
-
-      def not_equal(other)
-        other = other.to_node
-        other.is_a?(AST::Time) ? Boolean.new(value != other.value) : super
       end
     end
   end
